Support Allowed Paths from Environment Variable
This extends the functionality that's introduced with the `:without`
option so that these paths can be defined via a regex that's placed in
an environment variable.

If both `:without` and `:without_env` are defined, then the value in the
environment variable will take precedence. This allows for you to define
a default set of allowed paths in the `:without` value, while also
allowing for the paths to be changed during a maintenance activity
without needing a code change with the environment variable defined in
`:without_env`.